+++ js/json-canon-fuzz/src/bin.js
@@ -1,20 +1,31 @@
 #!/usr/bin/env node
 
+const { createWriteStream } = require('fs')
+const { join } = require('path')
+const { pipeline } = require('readable-stream')
+
 const fuzz = require('./')
 
 const args = process.argv.slice(2)
 
 const type = args[0]
-const numLines = parseInt(args[1])
+const numLines = args[1] ? parseInt(args[1]) : Infinity
 const outputFilePath = args[2]
 
-if (fuzz[type] == null || numLines === NaN) {
+if (fuzz[type] === undefined || Number.isNaN(numLines)) {
   usage()
 
   process.exit(1)
 }
 
-fuzz[type](numLines, outputFilePath)
+const getFuzzStream = fuzz[type]
+
+const fuzzStream = getFuzzStream({ numLines, outputFilePath })
+const outputStream = getOutputStream(outputFilePath)
+
+pipeline(fuzzStream, outputStream, function onDone(err) {
+  if (err) throw err
+})
 
 function usage() {
   console.log('Usage')
@@ -28,3 +39,12 @@ function usage() {
   console.log('Examples')
   console.log('  $ json-canon-fuzz json 100000')
 }
+
+function getOutputStream(outputFilePath) {
+  if (outputFilePath) {
+    const outputFileFullPath = join(process.cwd(), outputFilePath)
+    return createWriteStream(outputFileFullPath, { encoding: 'utf8' })
+  }
+  return process.stdout
+}
+
